user id type converted

+++ core/utils/jwt.go
@@ -9,7 +9,7 @@ import (
 
 func GenerateToken(user *models.User) (string, error) {
 	return jwt.NewWithClaims(jwt.SigningMethodHS256, jwt.MapClaims{
-		"user_id": int64(user.ID),
+		"user_id": user.ID,
 		"exp":     time.Now().Add(time.Hour * 2400000).Unix(),
 	}).SignedString([]byte(os.Getenv("JWT_SECRET")))
 }
